+++ sdk/ai/azure-ai-projects/CHANGELOG.md
@@ -4,9 +4,8 @@
 
 ### Bugs Fixed
 
-* Fix calls to methods `.inference.get_chat_completions_client()` and `.inference.get_embeddings_client()`, when Entra ID authentication
-is used.
-* Fixed bugs that were causing issues with tracing agent asynchronous functionality
+* Fix a bug in the `.inference` operations when Entra ID authentication is used by the default connection.
+* Fixed bugs that were causing issues with tracing in agent operations, when using the asynchronous client.
 * Fix a bug causing warning about unclosed session, shown when using asynchronous credentials to create agent.
 * Fix a bug that would cause agent function tool related function names and parameters to be included in traces even when content recording is not enabled